Sense function fixed / Secondary Main file added
Sense function fixed. Now the player won't know what object is in his vicinity, but will hear, smell, or be warned by the pokedex if a Pokemon, center, mart or trainer is nearby. This only affects the interaction with the prolog and have no impact on the "human gameplay".

Another addition was the main_no_prolog file. If you want to run the game without prolog interaction then run this file instead of the usual main.py